+++ src/cc/include/swcdb/fs/Local/FileSystem.h
@@ -42,11 +42,11 @@ class FileSystemLocal: public FileSystem {
                         const std::string &to) override;
   
   void write(int &err, SmartFd::Ptr &smartfd,
-             int32_t replication, int64_t blksz, 
+             uint8_t replication, int64_t blksz, 
              StaticBuffer &buffer);
 
   void create(int &err, SmartFd::Ptr &smartfd, 
-              int32_t bufsz, int32_t replication, int64_t blksz) override;
+              int32_t bufsz, uint8_t replication, int64_t blksz) override;
 
   void open(int &err, SmartFd::Ptr &smartfd, int32_t bufsz=0) override;
